Organize all utilities under `datalad_next.utils`
Ping #160
