+++ datalad_next/patches/push_optimize.py
@@ -3,10 +3,12 @@ import logging
 import re
 
 import datalad.core.distributed.push as mod_push
-from datalad_next.log import log_progress
 from datalad.runner.exception import CommandError
 from datalad.support.annexrepo import AnnexRepo
-from datalad_next.utils import ensure_list
+from datalad_next.utils import (
+    ensure_list,
+    log_progress,
+)
 from datalad_next.dataset import Dataset
