+++ datalad_next/url_operations/file.py
@@ -12,7 +12,7 @@ from urllib import (
     parse,
 )
 
-from datalad_next.consts import COPY_BUFSIZE
+from datalad_next.utils.consts import COPY_BUFSIZE
 
 from . import (
     UrlOperations,
