+++ datalad_next/utils/__init__.py
@@ -23,6 +23,10 @@ from datalad.utils import (
 from datalad.distribution.utils import _yield_ds_w_matching_siblings
 from datalad.support.external_versions import external_versions
 
+from .credman import CredentialManager
+from .log import log_progress
+
+
 lgr = logging.getLogger('datalad.utils')
 
 
@@ -70,7 +74,7 @@ def get_specialremote_credential_properties(params):
     # no other way to do this specifically for each supported remote type
     remote_type = params.get('type')
     if remote_type == 'webdav':
-        from datalad_next.http_helpers import get_auth_realm
+        from .http_helpers import get_auth_realm
         from datalad_next.url_operations.http import HttpUrlOperations
         url = params.get('url')
         if not url:
